load diff graphs.
